grcpPydenticSqlAlchemyAlembic/src/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
black==24.1.1
click==8.1.7
Faker==22.6.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
install==1.3.5
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
protobuf==4.25.2
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
